requirements.txt
# Core dependencies
requests>=2.31.0
python-dotenv>=1.0.0
pydantic>=2.5.0
numpy>=1.24.0
pandas>=2.0.0

# Reinforcement Learning
gymnasium>=0.29.0

# Deep Learning
torch>=2.0.0

# Visualization
matplotlib>=3.7.0

# Utilities
rich>=13.0.0
loguru>=0.7.0

# Web/API
streamlit>=1.28.0
fastapi>=0.104.0
uvicorn>=0.24.0

# Configuration
pyyaml>=6.0

# Cryptography (for Kalshi API signing)
cryptography>=41.0.0

# Scientific computing (for pricing models)
scipy>=1.11.0

# Testing
pytest>=7.4.0

